+++ examples/simple/lib/simple.ex
@@ -22,8 +22,8 @@ end
 defmodule Repo do
   use Ecto.Repo, adapter: Ecto.Adapters.Postgres
 
-  def url do
-    "ecto://postgres:postgres@localhost/ecto_simple"
+  def conf do
+    parse_url "ecto://postgres:postgres@localhost/ecto_simple"
   end
 
   def priv do