+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   * `Ecto.Binary[]` is no longer used to wrap binary values. Instead always use `binary/1` in queries
   * `:list` type changed name to `:array`. Need to specify inner type for arrays in entity fields
   * Literal lists no longer supported in queries. Need to specify inner type; use `array(list, ^:integer)` instead
+  * Remove `url/0` for configuration of repos in favor of `conf/0` in conjunction with `parse_url/1`
 
 # v0.0.1